Avoid signed/unsigned mismatch warning on DragonflyBSD.
For some reason, AT_FDCWD is defined as unsigned int on DragonflyBSD,
which causes warnings when int fds are compared against it.

+++ src/operations.cpp
@@ -1122,7 +1122,7 @@ uintmax_t remove_all_impl
 #if defined(BOOST_FILESYSTEM_HAS_FDOPENDIR_NOFOLLOW) && defined(BOOST_FILESYSTEM_HAS_POSIX_AT_APIS)
     fs::path filename;
     const fs::path* remove_path = &p;
-    if (parentdir_fd != AT_FDCWD)
+    if (parentdir_fd != static_cast< int >(AT_FDCWD))
     {
         filename = path_algorithms::filename_v4(p);
         remove_path = &filename;
